add profiling script
